Auto stash before rebase of "refs/heads/feature/refactor"
