+++ src/router/rotas/usuarios.go
@@ -49,29 +49,4 @@ var rotasUsuarios = []Rota{
 		Funcao:             controllers.AtualizarSenha,
 		RequerAutenticacao: true,
 	},
-
-	{
-		URI:                "/",
-		Metodo:             http.MethodGet,
-		Funcao:             controllers.CarregarTelaDeLogin,
-		RequerAutenticacao: false,
-	},
-	{
-		URI:                "/login",
-		Metodo:             http.MethodGet,
-		Funcao:             controllers.CarregarTelaDeLogin,
-		RequerAutenticacao: false,
-	},
-	{
-		URI:                "/login",
-		Metodo:             http.MethodPost,
-		Funcao:             controllers.FazerLogin,
-		RequerAutenticacao: false,
-	},
-	{
-		URI:                "/deslogar",
-		Metodo:             http.MethodGet,
-		Funcao:             controllers.Deslogar,
-		RequerAutenticacao: true,
-	},
 }
